+++ api/src/jobs/Index.test.ts
@@ -19,6 +19,7 @@ describe("Index", () => {
         let needleResponse: NeedleResponse;
         let consoleErrorSpy;
         let consoleLogSpy;
+        let sleepSpy;
         let unlockPidSpy;
         let queueSpy;
 
@@ -29,6 +30,7 @@ describe("Index", () => {
             indexer = {
                 deletePid: jest.fn(),
                 indexPid: jest.fn(),
+                getLastIndexResults: jest.fn(),
             };
             job = {
                 id: "1",
@@ -38,10 +40,12 @@ describe("Index", () => {
                 },
             } as Job;
             jest.spyOn(SolrIndexer, "getInstance").mockReturnValue(indexer);
+            indexer.getLastIndexResults.mockReturnValue({ fedora_parent_id_str_mv: ["foo:1"] });
             queueSpy = jest.spyOn(QueueManager.getInstance(), "getActiveIndexJobsForPid").mockResolvedValue([job]);
             consoleErrorSpy = jest.spyOn(console, "error").mockImplementation(jest.fn());
             consoleLogSpy = jest.spyOn(console, "log").mockImplementation(jest.fn());
             unlockPidSpy = jest.spyOn(SolrCache.getInstance(), "unlockPidIfEnabled").mockImplementation(jest.fn());
+            sleepSpy = jest.spyOn(index, "sleep").mockImplementation(jest.fn());
         });
 
         afterEach(() => {
@@ -104,6 +108,26 @@ describe("Index", () => {
             expect(unlockPidSpy).toHaveBeenCalledWith("vudl:123", "index");
         });
 
+        it("retries indexing the pid if parents are missing unexpectedly", async () => {
+            job.data.action = "index";
+            jest.spyOn(indexer, "indexPid").mockResolvedValue(needleResponse);
+            indexer.getLastIndexResults.mockReturnValueOnce({ fedora_parent_id_str_mv: [] });
+
+            await index.run(job);
+
+            expect(consoleErrorSpy).toHaveBeenCalledTimes(2);
+            expect(consoleErrorSpy).toHaveBeenNthCalledWith(
+                1,
+                new Error("vudl:123 has no parents and is not a configured top-level pid"),
+            );
+            expect(consoleErrorSpy).toHaveBeenNthCalledWith(2, "Retrying...");
+            expect(sleepSpy).toHaveBeenCalledTimes(1);
+            expect(consoleLogSpy).toHaveBeenCalledTimes(1);
+            expect(consoleLogSpy).toHaveBeenCalledWith("Indexing...", { action: "index", pid: "vudl:123" });
+            expect(indexer.indexPid).toHaveBeenCalledWith(job.data.pid);
+            expect(unlockPidSpy).toHaveBeenCalledWith("vudl:123", "index");
+        });
+
         it("throws an error when action does not exist", async () => {
             job.data.action = "fake action";
 
@@ -115,7 +139,6 @@ describe("Index", () => {
             const badJob = JSON.parse(JSON.stringify(job));
             job.id = "2";
             queueSpy.mockResolvedValue([badJob, job]);
-            const sleepSpy = jest.spyOn(index, "sleep").mockImplementation(jest.fn());
             await expect(index.run(job)).rejects.toThrow(/Exceeded retries waiting for queue to clear/);
             expect(sleepSpy).toHaveBeenCalledTimes(60);
             expect(sleepSpy).toHaveBeenCalledWith(1000);
@@ -124,7 +147,6 @@ describe("Index", () => {
         it("throws an error for statusCode not 200", async () => {
             needleResponse.statusCode = 404;
             jest.spyOn(indexer, "deletePid").mockResolvedValue(needleResponse);
-            const sleepSpy = jest.spyOn(index, "sleep").mockImplementation(jest.fn());
             await expect(index.run(job)).rejects.toThrow(/Problem performing/);
 
             expect(consoleErrorSpy).toHaveBeenCalledTimes(5);